+++ backend/docs/source/app.db.crud.rst
@@ -1,0 +1,29 @@
+app.db.crud package
+===================
+
+Submodules
+----------
+
+app.db.crud.topic\_crud module
+------------------------------
+
+.. automodule:: app.db.crud.topic_crud
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+app.db.crud.user\_crud module
+-----------------------------
+
+.. automodule:: app.db.crud.user_crud
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: app.db.crud
+   :members:
+   :undoc-members:
+   :show-inheritance:
