+++ backend/docs/source/conf.py
@@ -35,9 +35,9 @@ extensions = [
     "sphinx.ext.todo",
     "sphinx.ext.napoleon",
     "sphinx_rtd_theme",
-    'sphinx_autodoc_annotation',
-    'sphinx_autodoc_typehints',
-    'sphinx_git',
+    "sphinx_autodoc_annotation",
+    "sphinx_autodoc_typehints",
+    "sphinx_git",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
